refactor: move locate_device and force_provision to device category
These two tools use /cmd/devmgr and work on any UniFi device (switches,
APs, gateways), not just switches. Move them to the device category
alongside reboot_device, adopt_device, and upgrade_device where the
permission scope (devices) matches correctly.

- Move manager methods from SwitchManager to DeviceManager
- Move tool functions from tools/switch.py to tools/devices.py
- Change permission_category from "switch" to "devices"
- Update tool descriptions to note they work on any device
- Move tests to new test_device_commands.py
- Fix docs count (19 → 15 switch, 8 → 10 devices)
- Regenerate manifest (126 tools, same total)
