+++ apps/network/src/unifi_network_mcp/managers/switch_manager.py
@@ -346,50 +346,3 @@ class SwitchManager:
         await self._connection.request(api_request)
         return True
 
-    async def locate_device(self, device_mac: str, enabled: bool = True) -> bool:
-        """Enable or disable device locate (LED blinking).
-
-        Args:
-            device_mac: MAC address of the device.
-            enabled: True to start blinking, False to stop.
-
-        Returns:
-            True on success.
-
-        Raises:
-            Exception: On API error.
-        """
-        if not await self._connection.ensure_connected():
-            raise ConnectionError("Not connected to controller")
-
-        cmd = "set-locate" if enabled else "unset-locate"
-        api_request = ApiRequest(
-            method="post",
-            path="/cmd/devmgr",
-            data={"cmd": cmd, "mac": device_mac},
-        )
-        await self._connection.request(api_request)
-        return True
-
-    async def force_provision(self, device_mac: str) -> bool:
-        """Force re-provision a device.
-
-        Args:
-            device_mac: MAC address of the device.
-
-        Returns:
-            True on success.
-
-        Raises:
-            Exception: On API error.
-        """
-        if not await self._connection.ensure_connected():
-            raise ConnectionError("Not connected to controller")
-
-        api_request = ApiRequest(
-            method="post",
-            path="/cmd/devmgr",
-            data={"cmd": "force-provision", "mac": device_mac},
-        )
-        await self._connection.request(api_request)
-        return True
