+++ apps/network/src/unifi_network_mcp/tools/switch.py
@@ -586,77 +586,3 @@ async def set_jumbo_frames(
     except Exception as e:
         logger.error("Error setting jumbo frames on %s: %s", device_mac, e, exc_info=True)
         return {"success": False, "error": f"Failed to set jumbo frames on {device_mac}: {e}"}
-
-
-# ---- Device Management Commands ----
-
-
-@server.tool(
-    name="unifi_locate_device",
-    description="Toggle device locate mode (LED blinking) to physically identify a switch. "
-    "Requires confirmation.",
-    permission_category="switch",
-    permission_action="update",
-    annotations=ToolAnnotations(readOnlyHint=False, destructiveHint=False, idempotentHint=True, openWorldHint=False),
-)
-async def locate_device(
-    device_mac: Annotated[str, Field(description="MAC address of the device")],
-    enabled: Annotated[bool, Field(description="True to start blinking, False to stop")],
-    confirm: Annotated[
-        bool,
-        Field(description="When true, toggles locate mode. When false (default), returns a preview"),
-    ] = False,
-) -> Dict[str, Any]:
-    """Toggles device locate mode (LED blinking)."""
-    if not confirm:
-        action = "start" if enabled else "stop"
-        return create_preview(
-            resource_type="device_locate",
-            resource_data={"device_mac": device_mac, "enabled": enabled},
-            resource_name=device_mac,
-            warnings=[f"Will {action} LED blinking on device {device_mac}."],
-        )
-
-    try:
-        success = await switch_manager.locate_device(device_mac, enabled)
-        if success:
-            state = "enabled" if enabled else "disabled"
-            return {"success": True, "message": f"Locate mode {state} on device '{device_mac}'."}
-        return {"success": False, "error": f"Failed to set locate mode on '{device_mac}'."}
-    except Exception as e:
-        logger.error("Error setting locate mode on %s: %s", device_mac, e, exc_info=True)
-        return {"success": False, "error": f"Failed to set locate mode on {device_mac}: {e}"}
-
-
-@server.tool(
-    name="unifi_force_provision_device",
-    description="Force re-provision a device, pushing the current configuration from the "
-    "controller to the device. Useful after manual changes or to resolve config drift. "
-    "Requires confirmation.",
-    permission_category="switch",
-    permission_action="update",
-    annotations=ToolAnnotations(readOnlyHint=False, destructiveHint=False, idempotentHint=True, openWorldHint=False),
-)
-async def force_provision_device(
-    device_mac: Annotated[str, Field(description="MAC address of the device")],
-    confirm: Annotated[
-        bool,
-        Field(description="When true, force provisions the device. When false (default), returns a preview"),
-    ] = False,
-) -> Dict[str, Any]:
-    """Force re-provisions a device."""
-    if not confirm:
-        return create_preview(
-            resource_type="force_provision",
-            resource_data={"device_mac": device_mac},
-            resource_name=device_mac,
-        )
-
-    try:
-        success = await switch_manager.force_provision(device_mac)
-        if success:
-            return {"success": True, "message": f"Force provision initiated for device '{device_mac}'."}
-        return {"success": False, "error": f"Failed to force provision device '{device_mac}'."}
-    except Exception as e:
-        logger.error("Error force provisioning %s: %s", device_mac, e, exc_info=True)
-        return {"success": False, "error": f"Failed to force provision device {device_mac}: {e}"}
